app: fix port id type
Fixes: f8244c6399d9 ("ethdev: increase port id range")

+++ app/pdump/main.c
@@ -131,7 +131,7 @@ struct pdump_stats {
 
 struct pdump_tuples {
 	/* cli params */
-	uint8_t port;
+	uint16_t port;
 	char *device_id;
 	uint16_t queue;
 	char rx_dev[TX_STREAM_SIZE];
